# inventory/docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "This is a sample API Go server with Zap logging",
        "title": "API Go Service",
        "contact": {
            "name": "API Support",
            "url": "http://www.example.com/support",
            "email": "support@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8083",
    "basePath": "/api/v1",
    "paths": {
        "/inventory/count/": {
            "patch": {
                "description": "Возвращает сумму стоимости указанных товаров с учетом скидок",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Inventory"
                ],
                "summary": "Подсчет стоимости",
                "parameters": [
                    {
                        "description": "Список товаров для подсчета",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/structs.NewInventory"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Итоговая стоимость",
                        "schema": {
                            "$ref": "#/definitions/structs.InventorySwagger5"
                        }
                    },
                    "400": {
                        "description": "Неверно введены данные"
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера"
                    }
                }
            }
        },
        "/inventory/discount/": {
            "patch": {
                "description": "Устанавливает скидку на указанные товары на складе",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Inventory"
                ],
                "summary": "Установить скидку",
                "parameters": [
                    {
                        "description": "Данные для скидки",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/structs.NewInventoryDiscount"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Скидка успешно применена"
                    },
                    "400": {
                        "description": "Неверные входные данные"
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера"
                    }
                }
            }
        },
        "/inventory/goods/": {
            "get": {
                "description": "Возвращает список товаров на указанном складе с пагинацией",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Inventory"
                ],
                "summary": "Список товаров",
                "parameters": [
                    {
                        "description": "Параметры пагинации",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/structs.WarehousePagination"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список товаров",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/structs.ListByWarehouse"
                            }
                        }
                    },
                    "400": {
                        "description": "Неверные входные данные"
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера"
                    }
                }
            }
        },
        "/inventory/price/": {
            "post": {
                "description": "Устанавливает цену для конкретного товара на указанном складе",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Inventory"
                ],
                "summary": "Установить цену товара",
                "parameters": [
                    {
                        "description": "Данные установки цены",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/structs.InventorySwagger1"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Цена успешно установлена"
                    },
                    "400": {
                        "description": "Неверные входные данные"
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера"
                    }
                }
            }
        },
        "/inventory/product/": {
            "get": {
                "description": "Возвращает полную информацию о товаре на складе",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Inventory"
                ],
                "summary": "Информация о товаре",
                "parameters": [
                    {
                        "description": "Идентификаторы склада и товара",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/structs.InventorySwagger4"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Полная информация о товаре",
                        "schema": {
                            "$ref": "#/definitions/structs.AllInformationAboutTheProduct"
                        }
                    },
                    "400": {
                        "description": "Неверные входные данные"
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера"
                    }
                }
            }
        },
        "/inventory/purchase/": {
            "delete": {
                "description": "Обрабатывает покупку товаров и уменьшает их количество на складе",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Inventory"
                ],
                "summary": "Покупка товаров",
                "parameters": [
                    {
                        "description": "Данные о покупке",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/structs.NewInventory"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Покупка успешно обработана"
                    },
                    "400": {
                        "description": "Неверные входные данные или недостаточно товара"
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера"
                    }
                }
            }
        },
        "/inventory/updateQuantity/": {
            "patch": {
                "description": "Увеличивает количество товара на складе",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Inventory"
                ],
                "summary": "Обновить количество товара",
                "parameters": [
                    {
                        "description": "Данные обновления количества",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/structs.InventorySwagger2"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Количество успешно обновлено"
                    },
                    "400": {
                        "description": "Неверные входные данные"
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера"
                    }
                }
            }
        }
    },
    "definitions": {
        "structs.AllInformationAboutTheProduct": {
            "type": "object",
            "properties": {
                "barcode": {
                    "type": "string"
                },
                "characteristics": {
                    "type": "array",
                    "items": {
                        "type": "object",
                        "additionalProperties": {
                            "type": "string"
                        }
                    }
                },
                "description": {
                    "type": "string"
                },
                "discount": {
                    "type": "number"
                },
                "name": {
                    "type": "string"
                },
                "price": {
                    "type": "number"
                },
                "productId": {
                    "type": "string"
                },
                "quantity": {
                    "type": "integer"
                },
                "weight": {
                    "type": "number"
                }
            }
        },
        "structs.InventorySwagger1": {
            "type": "object",
            "properties": {
                "price": {
                    "type": "number",
                    "example": 999.99
                },
                "productId": {
                    "type": "string",
                    "example": "36c6ecf5-1e13-4cef-a591-bc8368fa0e60"
                },
                "warehouseId": {
                    "type": "string",
                    "example": "9be99fa0-2cd1-4c4e-863d-5115095fbd09"
                }
            }
        },
        "structs.InventorySwagger2": {
            "type": "object",
            "properties": {
                "productId": {
                    "type": "string",
                    "example": "36c6ecf5-1e13-4cef-a591-bc8368fa0e60"
                },
                "quantity": {
                    "type": "integer",
                    "example": 5
                },
                "warehouseId": {
                    "type": "string",
                    "example": "9be99fa0-2cd1-4c4e-863d-5115095fbd09"
                }
            }
        },
        "structs.InventorySwagger4": {
            "type": "object",
            "properties": {
                "productId": {
                    "type": "string",
                    "example": "36c6ecf5-1e13-4cef-a591-bc8368fa0e60"
                },
                "warehouseId": {
                    "type": "string",
                    "example": "9be99fa0-2cd1-4c4e-863d-5115095fbd09"
                }
            }
        },
        "structs.InventorySwagger5": {
            "type": "object",
            "properties": {
                "sum": {
                    "type": "number"
                }
            }
        },
        "structs.ListByWarehouse": {
            "type": "object",
            "properties": {
                "discount": {
                    "type": "number"
                },
                "name": {
                    "type": "string"
                },
                "price": {
                    "type": "number"
                },
                "productId": {
                    "type": "string"
                }
            }
        },
        "structs.NewInventory": {
            "description": "Данные для создания или обновления инвентарной записи",
            "type": "object",
            "properties": {
                "product": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/structs.ProductInventory2"
                    }
                },
                "warehouseId": {
                    "type": "string",
                    "example": "9be99fa0-2cd1-4c4e-863d-5115095fbd09"
                }
            }
        },
        "structs.NewInventoryDiscount": {
            "description": "Данные для установки скидки на товары",
            "type": "object",
            "properties": {
                "discount": {
                    "description": "Размер скидки (0.2 = 20%)",
                    "type": "number",
                    "example": 0.2
                },
                "productId": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "881670ed-8195-43d6-b678-3741bd3289ae",
                        "25f8f0ba-75f0-4caa-8648-27f6be3254b3"
                    ]
                },
                "warehouseId": {
                    "type": "string",
                    "example": "9be99fa0-2cd1-4c4e-863d-5115095fbd09"
                }
            }
        },
        "structs.ProductInventory2": {
            "description": "Упрощенная структура товара для операций",
            "type": "object",
            "properties": {
                "productId": {
                    "type": "string",
                    "example": "881670ed-8195-43d6-b678-3741bd3289ae"
                },
                "quantity": {
                    "type": "integer",
                    "example": 5
                }
            }
        },
        "structs.WarehousePagination": {
            "description": "Параметры пагинации для списка товаров на складе",
            "type": "object",
            "properties": {
                "limit": {
                    "description": "Количество элементов на странице",
                    "type": "integer",
                    "example": 10
                },
                "offset": {
                    "description": "Смещение (номер страницы * limit)",
                    "type": "integer",
                    "example": 0
                },
                "warehouseId": {
                    "type": "string",
                    "example": "9be99fa0-2cd1-4c4e-863d-5115095fbd09"
                }
            }
        }
    }
}